+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * register CSS properties so you can use CSS Transition on terminal colors [#808](https://github.com/jcubic/jquery.terminal/issues/808)
 * add terminal::blur alias [#813](https://github.com/jcubic/jquery.terminal/issues/813)
 * allow to change scrollbar color with CSS property
+* add `process_formatting` static helper function
+* add `FormattingCanvasRenderer` for color animation [#819](https://github.com/jcubic/jquery.terminal/issues/819)
 ### Bugfix
 * fix up/down arrows when moving inside command that has emoji [#608](https://github.com/jcubic/jquery.terminal/issues/608)
 * fix pausing when loading images [#807](https://github.com/jcubic/jquery.terminal/issues/807)